+++ shiny/experimental/ui_x/__init__.py
@@ -1,10 +1,11 @@
 from ._sidebar import layout_sidebar as layout_sidebar_x, sidebar as sidebar_x
 from ._page import page_fillable as page_fillable_x
-from ._card import (
+from ._card_item import (
+    as_card_items as as_card_items_x,
     as_card_item as as_card_item_x,
-    card as card_x,
     is_card_item as is_card_item_x,
 )
+from ._card import card as card_x
 from ._layout import layout_column_wrap as layout_column_wrap_x
 from ._valuebox import value_box as value_box_x
 
@@ -16,6 +17,7 @@ __all__ = (
     "page_fillable_x",
     # Card
     "as_card_item_x",
+    "as_card_items_x",
     "card_x",
     "is_card_item_x",
     # Layout
