+++ shiny/experimental/ui_x/_card_full_screen.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+from htmltools import HTML, Tag, tags
+
+from ._htmldeps import card_full_screen_dep
+
+
+def full_screen_toggle() -> Tag:
+    return tags.span(
+        {"class_": "bslib-full-screen-enter"},
+        {"class_": "badge rounded-pill bg-dark"},
+        {"data-bs-toggle": "tooltip"},
+        {"data-bs-placement": "bottom"},
+        {"title": "Expand"},
+        full_screen_toggle_icon(),
+        card_full_screen_dep(),
+    )
+
+
+# via bsicons::bs_icon("arrows-fullscreen")
+def full_screen_toggle_icon() -> HTML:
+    return HTML(
+        '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" class="bi bi-arrows-fullscreen " style="height:1em;width:1em;fill:currentColor;" aria-hidden="true" role="img" ><path fill-rule="evenodd" d="M5.828 10.172a.5.5 0 0 0-.707 0l-4.096 4.096V11.5a.5.5 0 0 0-1 0v3.975a.5.5 0 0 0 .5.5H4.5a.5.5 0 0 0 0-1H1.732l4.096-4.096a.5.5 0 0 0 0-.707zm4.344 0a.5.5 0 0 1 .707 0l4.096 4.096V11.5a.5.5 0 1 1 1 0v3.975a.5.5 0 0 1-.5.5H11.5a.5.5 0 0 1 0-1h2.768l-4.096-4.096a.5.5 0 0 1 0-.707zm0-4.344a.5.5 0 0 0 .707 0l4.096-4.096V4.5a.5.5 0 1 0 1 0V.525a.5.5 0 0 0-.5-.5H11.5a.5.5 0 0 0 0 1h2.768l-4.096 4.096a.5.5 0 0 0 0 .707zm-4.344 0a.5.5 0 0 1-.707 0L1.025 1.732V4.5a.5.5 0 0 1-1 0V.525a.5.5 0 0 1 .5-.5H4.5a.5.5 0 0 1 0 1H1.732l4.096 4.096a.5.5 0 0 1 0 .707z"></path></svg>'
+    )
